photo

+++ app/Http/Controllers/ADMIN/BlogController.php
@@ -28,34 +28,34 @@ class BlogController extends Controller
         return view('admin.blog.create', compact('title', 'categories'));
     }
 
-    public function store(Request $request)
-    {
-        $data =[
-            'category_id' => $request->category_id,
-            'title' => $request->title,
-            'slug' => Str::slug($request->title),
-            'content' => $request->content,
-            'status' => $request->status,
-            'writer' => auth()->user()->name,
-        ];
+    // public function store(Request $request)
+    // {
+    //     $data =[
+    //         'category_id' => $request->category_id,
+    //         'title' => $request->title,
+    //         'slug' => Str::slug($request->title),
+    //         'content' => $request->content,
+    //         'status' => $request->status,
+    //         'writer' => auth()->user()->name,
+    //     ];
 
-        if ($request->hasFile('photo')) {
-            $photo = $request->file('photo')->store('blog', 'public');
-            $data['photo'] = $photo;
-        }
-        Blog::create($data);
+    //     if ($request->hasFile('photo')) {
+    //         $photo = $request->file('photo')->store('blog', 'public');
+    //         $data['photo'] = $photo;
+    //     }
+    //     Blog::create($data);
 
-        Alert::success('Success', 'Create New Blog Successfully!');
-        return redirect()->to('admin/blog');
-    }
+    //     Alert::success('Success', 'Create New Blog Successfully!');
+    //     return redirect()->to('admin/blog');
+    // }
 
-    public function edit($id)
-    {
-        $edit = Blog::find($id);
-        $categories = Categories::get();
-        $title = "Edit Blog";
-        return view('admin.blog.edit', compact('edit', 'title', 'categories'));
-    }
+    // public function edit($id)
+    // {
+    //     $edit = Blog::find($id);
+    //     $categories = Categories::get();
+    //     $title = "Edit Blog";
+    //     return view('admin.blog.edit', compact('edit', 'title', 'categories'));
+    // }
 
     public function update(Request $request, $id)
     {
@@ -80,12 +80,37 @@ class BlogController extends Controller
         Alert::success('Success', 'Update Blog Successfully!');
         return redirect()->to('admin/blog');
     }
+        public function update(Request $request, $id)
+    {
+        $update = Blog::find($id);
+
+        $data = [
+            'category_id' => $request->category_id,
+            'title' => $request->title,
+            'slug' => Str::slug($request->title),
+            'content' => $request->content,
+            'status' => $request->status,
+            'writter' => auth()->user()->name,
+        ];
+
+        if ($request->hasFile('photo')) {
+            if ($update->photo) {
+                File::delete(public_path('storage/' . $update->photo));
+            }
+            $photo = $request->file('photo')->store('blog', 'public');
+            $data['photo'] = $photo;
+        }
+
+        $update->update($data);
+        return redirect()->to('admin/blog');
+    }
 
     public function destroy($id)
     {
-        $delete= Blog::find($id)->delete();
-        File::delete(public_path('storage/' . $update->photo));
-        Alert::success('Success', 'Delete Blog Successfully!');
+        $delete = Blog::find($id);
+        $delete->delete();
+        File::delete(public_path('storage/' . $delete->photo));
+        Alert::success('Success', 'Delete success!');
         return redirect()->to('admin/blog');
     }
 }
